feat(Crafty): add boost temp, brightness, and device info controls
- Implement set_boost_temperature (u16 LE to writeBoostTemp, clamped 0-30°C)
- Implement set_brightness (u16 LE to LED brightness characteristic)
- Implement get_device_info (reads firmware versions, use hours/minutes)
- Fix clippy warning: unnecessary cast in VolcanoHybrid::set_brightness

+++ src/protocol/volcano.rs
@@ -214,7 +214,7 @@ impl VaporizerControl for VolcanoHybrid {
 
     async fn set_brightness(&self, value: u16) -> Result<(), StorzError> {
         let ch = self.characteristic(VOLCANO_BRIGHTNESS).await?;
-        let raw = (value as u16).to_le_bytes();
+        let raw = value.to_le_bytes();
         self.peripheral
             .write(&ch, &raw, WriteType::WithoutResponse)
             .await?;
